requirements.txt
# EnableBot - Production Dependencies

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.2

# Data Validation
pydantic>=2.8.0
pydantic-settings>=2.0.0

# Database
asyncpg>=0.29.0

# Encryption
cryptography>=41.0.7

# Environment
python-dotenv>=1.0.0

# Templates
jinja2>=3.1.0

# File Processing
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.0

# AI Services
openai>=1.3.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1